fix AudioFile.Read
the loop in Read attempted to limit the data copied to the length of a
single chunk, but if the cursor wasn't at the start of a chunk then it
was possible to read beyond the bounds of the current chunk. if the next
chunk hadn't been loaded yet, this would return some garbage zero
values.

while we're here, simplify the read logic: copy already computes the
minimum between the two slice lengths and returns the number of bytes
copies, so there is no need to do that ourselves