fix an overread in AudioFile.Read
This fixes a race in Read: it attempted to limit the data copied to the
length of a single chunk, but if the cursor wasn't at the start of a
chunk then it was possible to read beyond the bounds of the current
chunk. If the next chunk hadn't been loaded yet, this would return some
garbage zero values.